DAYD-727 refactor フォルダ構成とファイル名を一部変更
